Add form validation to the search input
Now when the user enter a value in the search input, will check if its
empty, if starts with a number or if is shorter than 3 characters, if
any of that is true, an error message will be displayed.

We use the useState to save the search and error values, and the
useEffect to validate the form.