Search movies with the user input
Now then the user enter a value will search for movies that match that
search, as the user types, but also with the submit button, this takes
place in the  handleChange and handleSubmit functions calling the
getMovies function with the query state value for the handleSubmit and
the value of the input for the handleChange.

Some errors were fixed like not having an await on the response.json()
and the some parameters prepared to receive an object with a property of
search instead of query.